fix: change kirillic letter
