+++ packages/client/src/game/logic/circle/hitLogic.ts
@@ -1,15 +1,15 @@
-import Circle from "game/objects/сircle";
+import Circle from 'game/objects/circle';
 
 export function checkHit(
-  circles: Circle[],
-  x: number,
-  y: number
+    circles: Circle[],
+    x: number,
+    y: number
 ): { hit: boolean; circle?: Circle } {
-  for (const circle of circles) {
-    if (circle.containsPoint(x, y)) {
-      circle.pop();
-      return { hit: true, circle };
+    for (const circle of circles) {
+        if (circle.containsPoint(x, y)) {
+            circle.pop();
+            return { hit: true, circle };
+        }
     }
-  }
-  return { hit: false };
+    return { hit: false };
 }
